moving az_base64 to core private
