+++ sdk/core/core/CMakeLists.txt
@@ -64,6 +64,7 @@ else()
 endif()
 
 target_include_directories (az_core PUBLIC inc)
+target_include_directories (az_core PRIVATE src/private)
 
 # make sure that users can consume the project as a library.
 add_library (az::core ALIAS az_core)
@@ -79,6 +80,7 @@ add_executable (
     )
 
 target_link_libraries(az_core_test PRIVATE az_core)
+target_include_directories (az_core_test PRIVATE src/private)
 
 target_link_libraries(az_core PRIVATE CURL::libcurl)
