+++ sdk/core/core/src/private/az_base64.h
